--- map.py
@@ -1,28 +1,0 @@
-class Map:
-    def __init__():
-        """
-        Hier wird die Größe der Karte festgelegt.
-        """
-        pass
-
-    def ant_nest():
-        """
-        Nest wird zufällig auf der Karte platziert und enthält alle Ameisen.
-        Die Anzahl an Ameisen können mittels eines Inputs festgelegt werden.
-        """
-        pass
-
-    def food_source():
-        """
-        Die Nahrungsquellen werden ebenfalls zufällig auf der Karte platziert.
-        Man kann ein int mitangeben um die Anzahl an Nahrungsquellen festzulegen.
-        Außerdem wird die "Lebenszeit" der Nahrungsquelle bestimmt. Also wie viele Ameisen braucht es bis die Nahrungsquelle aufgebraucht ist.
-        """
-        pass
-
-    def obstacle():
-        """
-        Die Nahrungsquellen werden ebenfalls zufällig auf der Karte platziert.
-        Man kann ein int mitangeben um die Anzahl an Nahrungsquellen festzulegen.
-        """
-        pass
